Update RoleBinding manifests to ClusterRoleBindings
This is a temporary and probably sub-optimal fix for development
expedience. The RoleBinding wasn't allowing the default ServiceAccount
in the foghorn namespace to create GitEvents and Actions. This allows
the default foghorn SA admin access to those resources at the cluster
scope.

A better long-term solution would be to define a foghorn ServiceAccount
specifically for managing GitEvents and Actions.